requirements.txt
appnope==0.1.0
backcall==0.1.0
beautifulsoup4==4.6.0
boto3==1.7.44
botocore==1.10.44
cachetools==2.1.0
certifi==2018.4.16
chardet==3.0.4
click==6.7
decorator==4.3.0
docutils==0.14
Flask==1.0.2
google-api-core==1.2.1
google-auth==1.5.0
google-cloud-texttospeech==0.2.0
googleapis-common-protos==1.5.3
grpcio==1.12.1
idna==2.7
ipython==6.4.0
ipython-genutils==0.2.0
itsdangerous==0.24
jedi==0.12.0
Jinja2==2.10
jmespath==0.9.3
MarkupSafe==1.0
parso==0.2.1
pexpect==4.6.0
pickleshare==0.7.4
prompt-toolkit==1.0.15
protobuf==3.6.0
ptyprocess==0.6.0
pyasn1==0.4.3
pyasn1-modules==0.2.1
Pygments==2.2.0
python-dateutil==2.7.3
pyttsx3==2.7
pytz==2018.4
requests==2.19.1
rsa==3.4.2
s3transfer==0.1.13
simplegeneric==0.8.1
six==1.11.0
slackclient==1.2.1
traitlets==4.3.2
urllib3==1.23
wcwidth==0.1.7
websocket-client==0.48.0
Werkzeug==0.14.1
